split up extension.ts and changed how its compiled
