+++ package.json
@@ -61,10 +61,10 @@
   "repository": "https://github.com/lbauskar/kanban.git",
   "homepage": "/",
   "scripts": {
-    "vscode:prepublish": "(webpack --config webpack.config.js --mode=production) && tsc -p tsconfig.extension.json",
-    "build": "(webpack --config webpack.config.js --mode=development) && tsc -p tsconfig.extension.json",
+    "vscode:prepublish": "(webpack --config webpack.config.js --mode=production) && (webpack --config webpack-extension.config.js --mode=production)",
+    "build": "(webpack --config webpack.config.js --mode=development) && (webpack --config webpack-extension.config.js --mode=development)",
     "test": "jest --env=jsdom",
-    "package": "yarn vsce package"
+    "package": "vsce package"
   },
   "devDependencies": {
     "@testing-library/dom": "^8.1.0",
@@ -97,6 +97,7 @@
     "copy-webpack-plugin": "^10.2.4",
     "cuid": "^2.1.8",
     "just-clone": "^3.2.1",
+    "path-browserify": "^1.0.1",
     "react": "^17.0.2",
     "react-beautiful-dnd": "^13.1.0",
     "react-colorful": "^5.3.0",
