+++ tsconfig.json
@@ -22,5 +22,5 @@
         "jsx": "react-jsx"
     },
     "include": ["src"],
-    "exclude": ["src/extension.ts"]
+    "exclude": ["src/extension"]
 }
